+++ src/main/java/com/mju/mentoring/member/infrastructure/jwtmanager/JwtManager.java
@@ -1,4 +1,4 @@
-package com.mju.mentoring.member.infrastructure;
+package com.mju.mentoring.member.infrastructure.jwtmanager;
 
 public interface JwtManager {
